Factor logic for well-known types/fns into its own file
