+++ compiler/parser/parser.cpp
@@ -31,6 +31,7 @@
 #include "insertLineNumbers.h"
 #include "stringutil.h"
 #include "symbol.h"
+#include "wellknown.h"
 
 #include <cstdlib>
 
@@ -264,9 +265,6 @@ static void countTokensInCmdLineFiles() {
 *                                                                             *
 ************************************** | *************************************/
 
-static void gatherIteratorTags();
-static void gatherWellKnownTypes();
-static void gatherWellKnownFns();
 
 static void parseInternalModules() {
   if (fDocs == false || fDocsProcessUsedModules == true) {
@@ -282,220 +280,6 @@ static void parseInternalModules() {
   }
 }
 
-static void gatherIteratorTags() {
-  forv_Vec(TypeSymbol, ts, gTypeSymbols) {
-    if (strcmp(ts->name, iterKindTypename) == 0) {
-      if (EnumType* enumType = toEnumType(ts->type)) {
-        for_alist(expr, enumType->constants) {
-          if (DefExpr* def = toDefExpr(expr)) {
-            const char* name = def->sym->name;
-
-            if        (strcmp(name, iterKindLeaderTagname)     == 0) {
-              gLeaderTag     = def->sym;
-
-            } else if (strcmp(name, iterKindFollowerTagname)   == 0) {
-              gFollowerTag   = def->sym;
-
-            } else if (strcmp(name, iterKindStandaloneTagname) == 0) {
-              gStandaloneTag = def->sym;
-            }
-          }
-        }
-      }
-    }
-  }
-}
-
-// This structure and the following array provide a list of types that must be
-// defined in module code.  At this point, they are all classes.
-struct WellKnownType
-{
-  const char*     name;
-  AggregateType** type_;
-  bool            isClass;
-};
-
-// These types are a required part of the compiler/module interface.
-static WellKnownType sWellKnownTypes[] = {
-  { "_array",                &dtArray,            false },
-  { "_tuple",                &dtTuple,            false },
-  { "locale",                &dtLocale,           true  },
-  { "chpl_localeID_t",       &dtLocaleID,         false },
-  { "BaseArr",               &dtBaseArr,          true  },
-  { "BaseDom",               &dtBaseDom,          true  },
-  { "BaseDist",              &dtDist,             true  },
-  { "chpl_main_argument",    &dtMainArgument,     false },
-  { "chpl_comm_on_bundle_t", &dtOnBundleRecord,   false },
-  { "chpl_task_bundle_t",    &dtTaskBundleRecord, false },
-  { "Error",                 &dtError,            true  }
-};
-
-// Gather well-known types from among types known at this point.
-static void gatherWellKnownTypes() {
-  int nEntries = sizeof(sWellKnownTypes) / sizeof(sWellKnownTypes[0]);
-
-  // Harvest well-known types from among the global type definitions.
-  // We check before assigning to the well-known type dt<typename>,
-  // to ensure that it is null.  In that way we can flag duplicate
-  // definitions.
-  forv_Vec(TypeSymbol, ts, gTypeSymbols) {
-    for (int i = 0; i < nEntries; ++i) {
-      WellKnownType& wkt = sWellKnownTypes[i];
-
-      if (strcmp(ts->name, wkt.name) == 0) {
-        if (*wkt.type_ != NULL) {
-          USR_WARN(ts,
-                   "'%s' defined more than once in Chapel internal modules.",
-                   wkt.name);
-        }
-
-        INT_ASSERT(ts->type);
-
-        if (wkt.isClass == true && isClass(ts->type) == false) {
-          USR_FATAL_CONT(ts->type,
-                         "The '%s' type must be a class.",
-                         wkt.name);
-        }
-
-        *wkt.type_ = toAggregateType(ts->type);
-      }
-    }
-  }
-
-  if (fMinimalModules == false) {
-    // Make sure all well-known types are defined.
-    for (int i = 0; i < nEntries; ++i) {
-      WellKnownType& wkt = sWellKnownTypes[i];
-
-      if (*wkt.type_ == NULL) {
-        USR_FATAL_CONT("Type '%s' must be defined in the "
-                       "Chapel internal modules.",
-                       wkt.name);
-      }
-    }
-
-    USR_STOP();
-
-  } else {
-    if (dtString->symbol == NULL) {
-      // This means there was no declaration of the string type.
-      gAggregateTypes.remove(gAggregateTypes.index(dtString));
-
-      delete dtString;
-
-      dtString = NULL;
-    }
-  }
-}
-
-struct WellKnownFn
-{
-  const char* name;
-  FnSymbol**  fn;
-  Flag        flag;
-  FnSymbol*   lastNameMatchedFn;
-};
-
-// These functions are a required part of the compiler/module interface.
-// They should generally be marked export so that they are always
-// resolved.
-static WellKnownFn sWellKnownFns[] = {
-  {
-    "chpl_here_alloc",
-    &gChplHereAlloc,
-    FLAG_LOCALE_MODEL_ALLOC
-  },
-
-  {
-    "chpl_here_free",
-    &gChplHereFree,
-    FLAG_LOCALE_MODEL_FREE
-  },
-
-  {
-    "chpl_doDirectExecuteOn",
-    &gChplDoDirectExecuteOn,
-    FLAG_UNKNOWN
-  },
-
-  {
-    "_build_tuple",
-    &gBuildTupleType,
-    FLAG_BUILD_TUPLE_TYPE
-  },
-
-  {
-    "_build_tuple_noref",
-    &gBuildTupleTypeNoRef,
-    FLAG_BUILD_TUPLE_TYPE
-  },
-
-  {
-    "*",
-    &gBuildStarTupleType,
-    FLAG_BUILD_TUPLE_TYPE
-  },
-
-  {
-    "_build_star_tuple_noref",
-    &gBuildStarTupleTypeNoRef,
-    FLAG_BUILD_TUPLE_TYPE
-  },
-
-  {
-    "chpl_delete_error",
-    &gChplDeleteError,
-    FLAG_UNKNOWN
-  }
-};
-
-static void gatherWellKnownFns() {
-  int nEntries = sizeof(sWellKnownFns) / sizeof(sWellKnownFns[0]);
-
-  // Harvest well-known functions from among the global fn definitions.
-  // We check before assigning to the associated global to ensure that it
-  // is null.  In that way we can flag duplicate definitions.
-  forv_Vec(FnSymbol, fn, gFnSymbols) {
-    for (int i = 0; i < nEntries; ++i) {
-      WellKnownFn& wkfn = sWellKnownFns[i];
-
-      if (strcmp(fn->name, wkfn.name) == 0) {
-        wkfn.lastNameMatchedFn = fn;
-
-        if (wkfn.flag == FLAG_UNKNOWN || fn->hasFlag(wkfn.flag) == true) {
-          if (*wkfn.fn != NULL) {
-            USR_WARN(fn,
-                     "'%s' defined more than once in Chapel internal modules.",
-                     wkfn.name);
-          }
-
-          *wkfn.fn = fn;
-        }
-      }
-    }
-  }
-
-  if (fMinimalModules == false) {
-    for (int i = 0; i < nEntries; ++i) {
-      WellKnownFn& wkfn        = sWellKnownFns[i];
-      FnSymbol*    lastMatched = wkfn.lastNameMatchedFn;
-      FnSymbol*    fn          = *wkfn.fn;
-
-      if (lastMatched == NULL) {
-        USR_FATAL_CONT("Function '%s' must be defined in the "
-                       "Chapel internal modules.",
-                       wkfn.name);
-
-      } else if (fn == NULL) {
-        USR_FATAL_CONT(fn,
-                       "The '%s' function is missing a required flag.",
-                       wkfn.name);
-      }
-    }
-
-    USR_STOP();
-  }
-}
 
 /************************************* | **************************************
 *                                                                             *
